Add suggestions from the code review
