+++ commons/src/main/java/io/ballerina/stdlib/graphql/commons/types/DefaultDirective.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package io.ballerina.stdlib.graphql.commons.schema.types;
+package io.ballerina.stdlib.graphql.commons.types;
 
 import java.util.Arrays;
 import java.util.List;
