+++ commons/src/main/java/module-info.java
@@ -17,7 +17,6 @@
  */
 
 module io.ballerina.stdlib.graphql.commons {
-    exports io.ballerina.stdlib.graphql.commons;
-    exports io.ballerina.stdlib.graphql.commons.schema;
-    exports io.ballerina.stdlib.graphql.commons.schema.types;
+    exports io.ballerina.stdlib.graphql.commons.utils;
+    exports io.ballerina.stdlib.graphql.commons.types;
 }
