+++ native/src/main/java/io/ballerina/stdlib/graphql/runtime/engine/Engine.java
@@ -33,7 +33,7 @@ import io.ballerina.runtime.api.values.BArray;
 import io.ballerina.runtime.api.values.BError;
 import io.ballerina.runtime.api.values.BObject;
 import io.ballerina.runtime.api.values.BString;
-import io.ballerina.stdlib.graphql.commons.schema.types.Schema;
+import io.ballerina.stdlib.graphql.commons.types.Schema;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
