+++ native/src/main/java/io/ballerina/stdlib/graphql/runtime/engine/SchemaRecordGenerator.java
@@ -28,14 +28,14 @@ import io.ballerina.runtime.api.utils.StringUtils;
 import io.ballerina.runtime.api.values.BArray;
 import io.ballerina.runtime.api.values.BMap;
 import io.ballerina.runtime.api.values.BString;
-import io.ballerina.stdlib.graphql.commons.schema.types.Directive;
-import io.ballerina.stdlib.graphql.commons.schema.types.DirectiveLocation;
-import io.ballerina.stdlib.graphql.commons.schema.types.EnumValue;
-import io.ballerina.stdlib.graphql.commons.schema.types.Field;
-import io.ballerina.stdlib.graphql.commons.schema.types.InputValue;
-import io.ballerina.stdlib.graphql.commons.schema.types.Schema;
-import io.ballerina.stdlib.graphql.commons.schema.types.Type;
-import io.ballerina.stdlib.graphql.commons.schema.types.TypeKind;
+import io.ballerina.stdlib.graphql.commons.types.Directive;
+import io.ballerina.stdlib.graphql.commons.types.DirectiveLocation;
+import io.ballerina.stdlib.graphql.commons.types.EnumValue;
+import io.ballerina.stdlib.graphql.commons.types.Field;
+import io.ballerina.stdlib.graphql.commons.types.InputValue;
+import io.ballerina.stdlib.graphql.commons.types.Schema;
+import io.ballerina.stdlib.graphql.commons.types.Type;
+import io.ballerina.stdlib.graphql.commons.types.TypeKind;
 
 import java.util.ArrayList;
 import java.util.HashMap;
